+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## dispatio 1.1.0  not yet released
 
+* Focus on #call, alias #dispatch to it
+
 
 ## dispatio 1.0.0  released 2026-09-11